Move MXSerializer to unique namespace
 - prevents dependency version from loading
 - patch MXSerializer to skip duplicate namespaces at same depth
